chore: implement controllers
